Fix implicit cast.

+++ unizone/src/acronymclient.cpp
@@ -96,7 +96,7 @@ AcronymClient::SessionConnected(const String & /* sessionID */)
 		if (fHostPort != 80)
 		{
 			cmd << ":";
-			cmd << (long) fHostPort;
+			cmd << (int) fHostPort;
 		}
 		cmd << "\n\n";
 		ref()->AddString(PR_NAME_TEXT_LINE, cmd);
