+++ unizone/src/updateclient.cpp
@@ -73,7 +73,7 @@ UpdateClient::SessionConnected(const String & /* sessionID */)
 		if (fHostPort != 80)
 		{
 			cmd << ":";
-			cmd << (long) fHostPort;
+			cmd << (int) fHostPort;
 		}
 		cmd << "\n\n";
 		ref()->AddString(PR_NAME_TEXT_LINE, cmd);
